Keep Coding

+++ ShortVideo/src/main/java/com/lib/shortvideo/utils/FileUtil.java
@@ -1,8 +1,7 @@
 package com.lib.shortvideo.utils;
 
 import android.os.Environment;
-
-import com.lib.shortvideo.videoview.recorder.Constants;
+import android.text.TextUtils;
 
 import java.io.File;
 
@@ -29,26 +28,12 @@ public class FileUtil
      */
     public static boolean deleteFile(String filePath)
     {
+        if (TextUtils.isEmpty(filePath))
+            return false;
+
         File file = new File(filePath);
         if (file.exists())
             return file.delete();
         return true;
-    }
-
-    /**
-     * 创建文件名
-     *
-     * @param folder    父文件夹
-     * @param subfolder 子文件加
-     * @param uniqueId  文件名唯一识别码
-     */
-    public static String createFilePath(String folder, String subfolder, String uniqueId)
-    {
-        File dir = new File(Environment.getExternalStorageDirectory(), folder);
-        if (subfolder != null)
-            dir = new File(dir, subfolder);
-        dir.mkdirs();
-        String fileName = Constants.FILE_START_NAME + uniqueId + Constants.VIDEO_EXTENSION;
-        return new File(dir, fileName).getAbsolutePath();
     }
 }
